Double error node issue fix
